requirements.txt
Flask==2.3.3
twilio==9.6.3
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7